added main function to test

+++ read_hdf5.h
@@ -12,7 +12,7 @@
 
 typedef struct halo
 {
-  double mass;
+  float mass;
   float X;
   float Y;
   float Z;
